A whole lot of refactoring. The plugin is growing, I need to separate the views.
