+++ new-client/src/plugins/PropertyChecker/views/DigitalPlanChecker/DigitalPlanItem.js
