+++ new-client/src/plugins/PropertyChecker/views/LayerChecker/FeatureItem.js
