+++ new-client/src/plugins/PropertyChecker/views/LayerChecker/ReportDialog.js
